Resolving compiler warnings in Actions and Math module

+++ src/BabylonCpp/include/babylon/actions/directactions/play_animation_action.h
@@ -10,7 +10,7 @@ class BABYLON_SHARED_EXPORT PlayAnimationAction : public Action {
 
 public:
   PlayAnimationAction(unsigned int triggerOptions, IAnimatable* target,
-                      int from, int to, bool loop = false,
+                      float from, float to, bool loop = false,
                       Condition* condition = nullptr);
   ~PlayAnimationAction();
 
@@ -19,8 +19,8 @@ public:
   Json::object serialize(Json::object& parent) const override;
 
 public:
-  int from;
-  int to;
+  float from;
+  float to;
   bool loop;
 
 private:
